refactor: just translate the minimal example
