+++ Cargo.toml
@@ -79,7 +79,6 @@ members = [
     "examples/trace-transaction-cli/",
     "examples/polygon-p2p/",
     "examples/custom-inspector/",
-    "examples/exex/async-fn/",
     "examples/exex/minimal/",
     "examples/exex/op-bridge/",
     "testing/ef-tests/",
