--- examples/exex/async-fn/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "async-fn"
-version = "0.0.0"
-publish = false
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-reth.workspace = true
-reth-exex.workspace = true
-reth-node-api.workspace = true
-reth-node-core.workspace = true
-reth-node-ethereum.workspace = true
-reth-primitives.workspace = true
-reth-provider.workspace = true
-
-eyre.workspace = true
-tokio.workspace = true
-futures.workspace = true
